ai_women_safety_project/requirements.txt
Flask
Flask-Cors
mysql-connector-python
twilio
